+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 * Corner cases in `kwarray.stats_dict`
 * Issue in `group_indices` when given a special view array
+* Fixed numpy warnings
+* Fixed usage of `torch.complex32`
 
 
 ## Version 0.6.0 - Released 2022-03-04
@@ -21,7 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `util_robust` for robust normalization.
 * Add `return_index` to `unique_rows`
 
-### Changes
+### Fixed
 * Fixes to normalize
 * Remove support for Python < 3.6